# Optimizing siphash implementation