--- a/src/bench/crypto_hash.cpp
+++ b/src/bench/crypto_hash.cpp
@@ -85,8 +85,29 @@ static void SipHash_32b(benchmark::Bench& bench)
 {
     uint256 x;
     uint64_t k1 = 0;
-    bench.run([&] {
+    bench.batch(x.size()).unit("byte").run([&] {
         *((uint64_t*)x.begin()) = SipHashUint256(0, ++k1, x);
+    });
+}
+
+static void SipHash_3b(benchmark::Bench& bench)
+{
+    uint64_t hash = 0;
+    uint64_t k2 = 0;
+    std::vector<uint8_t> in(3, 0xee);
+    bench.batch(in.size()).unit("byte").run([&] {
+        hash = CSipHasher(hash, ++k2).Write(in.data(), in.size()).Finalize();
+    });
+}
+
+
+static void SipHash(benchmark::Bench& bench)
+{
+    uint64_t hash = 0;
+    uint64_t k2 = 0;
+    std::vector<uint8_t> in(BUFFER_SIZE, 0);
+    bench.batch(in.size()).unit("byte").run([&] {
+        hash = CSipHasher(hash, ++k2).Write(in.data(), in.size()).Finalize();
     });
 }
 
@@ -155,9 +176,11 @@ BENCHMARK(SHA1);
 BENCHMARK(SHA256);
 BENCHMARK(SHA512);
 BENCHMARK(SHA3_256_1M);
+BENCHMARK(SipHash);
 
 BENCHMARK(SHA256_32b);
 BENCHMARK(SipHash_32b);
+BENCHMARK(SipHash_3b);
 BENCHMARK(SHA256D64_1024);
 BENCHMARK(FastRandom_32bit);
 BENCHMARK(FastRandom_1bit);