@laanwj 2020-01-29
That's a very nice speed improvement!