@sipa 2020-01-29
Where do we use variable-length SipHash?