@emilengler 2020-01-29
I just tested it and the speed improvement is good. Also the formatting is much prettier now.
Will review the code now