@elichai 2020-01-29
> Where do we use variable-length SipHash?

A quick search shows:
1. `GCSFilter::HashToRange` BIP158 Compact block filters (blockfilters.h)
2. `RelayAddress` I think as a PRNG (net_processing.h).
3. `ByteVectorHash` used as the hash for the `std::unordered_set<Element, ByteVectorHash> ElementSet;` in `blockfilter.h`.
4. `CConnman::GetDeterministicRandomizer`  again some randomizer thing (net.cpp).