@Empact 2020-01-30
nit: You have a few whitespace irregularities - running `git-clang-format` generates this diff

```diff
diff --git a/src/bench/crypto_hash.cpp b/src/bench/crypto_hash.cpp
index 9eeb8da16..037260939 100644
--- a/src/bench/crypto_hash.cpp
+++ b/src/bench/crypto_hash.cpp
@@ -90,7 +90,7 @@ static void SipHash(benchmark::State& state)
 {
     uint64_t hash = 0;
     uint64_t k2 = 0;
-    std::vector<uint8_t> in(BUFFER_SIZE,0);
+    std::vector<uint8_t> in(BUFFER_SIZE, 0);
     while (state.KeepRunning())
         hash = CSipHasher(hash, ++k2).Write(in.data(), in.size()).Finalize();
 }
diff --git a/src/crypto/siphash.cpp b/src/crypto/siphash.cpp
index 0b62de998..cfc04c194 100644
--- a/src/crypto/siphash.cpp
+++ b/src/crypto/siphash.cpp
@@ -2,8 +2,8 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <crypto/siphash.h>
 #include <crypto/common.h>
+#include <crypto/siphash.h>
 
 #include <algorithm>
 
@@ -50,7 +50,8 @@ CSipHasher& CSipHasher::Write(uint64_t data)
 
 
 /// Load a uint64_t from 0 to 7 bytes.
-inline uint64_t ReadU64ByLenLE(const unsigned char* data, size_t len) {
+inline uint64_t ReadU64ByLenLE(const unsigned char* data, size_t len)
+{
     assert(len < 8);
     uint64_t out = 0;
     for (size_t i = 0; i < len; ++i) {
@@ -85,7 +86,7 @@ CSipHasher& CSipHasher::Write(const unsigned char* data, size_t size)
 
     auto i = needed;
     while (i < len - left) {
-        uint64_t mi = ReadLE64(data+i);
+        uint64_t mi = ReadLE64(data + i);
         v3 ^= mi;
         SIPROUND;
         SIPROUND;
diff --git a/src/crypto/siphash.h b/src/crypto/siphash.h
index 698f9c310..ddf192e8a 100644
--- a/src/crypto/siphash.h
+++ b/src/crypto/siphash.h
@@ -14,7 +14,7 @@ class CSipHasher
 {
 private:
     uint64_t v[4];
-    size_t count; // total amount of bytes inputted.
+    size_t count;  // total amount of bytes inputted.
     uint64_t tail; // bytes that weren't processed yet.
 
 public:
```