@jonatack 2020-02-23
> I decided to drop commit [2b32471](https://github.com/bitcoin/bitcoin/commit/2b32471b5b113c0a34e06896842a450dc7168454) because it's more controversial than I thought (there might be software that relies on the current formatting and there's #18011 coming up)

Right, will be easier to review as more focused; it was 2 PRs in one before.