@m2709 2020-02-24
El dom., 23 de febrero de 2020 11:27, Jon Atack <notifications@github.com>
escribió:

> I decided to drop commit 2b32471
> <https://github.com/bitcoin/bitcoin/commit/2b32471b5b113c0a34e06896842a450dc7168454>
> because it's more controversial than I thought (there might be software
> that relies on the current formatting and there's #18011
> <https://github.com/bitcoin/bitcoin/pull/18011> coming up)
>
> Right, will be easier to review as more focused; it was 2 PRs in one
> before.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/bitcoin/bitcoin/pull/18014?email_source=notifications&email_token=AOHIX5ATCPUKTZ3AX7F23VDREKBU5A5CNFSM4KMWPL3KYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEMV5DZQ#issuecomment-590074342>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AOHIX5DCBMU6TEISF3UMOBTREKBU5ANCNFSM4KMWPL3A>
> .
>