@elichai 2020-08-02
Updated benchmarks with the new benchmark library
Before:
```
$ sudo pyperf system tune
$ ./src/bench/bench_bitcoin -filter="SipHash|SipHash_3b|SipHash_32b"
|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|        2,981,360.00 |              335.42 |    0.0% |   23,000,202.00 |    7,146,000.00 |  3.219 |   2,125,014.00 |    0.0% |      0.03 | `SipHash`
|               39.51 |       25,311,174.47 |    0.1% |          237.03 |           94.78 |  2.501 |           3.00 |    0.1% |      0.00 | `SipHash_32b`
|               27.88 |       35,874,274.80 |    0.1% |          239.02 |           66.91 |  3.572 |          17.00 |    0.0% |      0.00 | `SipHash_3b`

```

After:
```
$ sudo pyperf system tune
$ ./src/bench/bench_bitcoin -filter="SipHash|SipHash_3b|SipHash_32b"
|               ns/op |                op/s |    err% |          ins/op |          cyc/op |    IPC |         bra/op |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|          676,474.00 |            1,478.25 |    0.0% |    4,250,201.00 |    1,623,700.00 |  2.618 |     125,015.00 |    0.0% |      0.01 | `SipHash`
|               39.55 |       25,282,631.04 |    0.2% |          237.03 |           94.86 |  2.499 |           3.00 |    0.2% |      0.00 | `SipHash_32b`
|               23.69 |       42,213,307.31 |    0.2% |          193.02 |           56.96 |  3.389 |          16.00 |    0.0% |      0.00 | `SipHash_3b`
```